pyproject.toml
[tool.poetry]
name = "market-screener"
version = "0.1.0"
description = "Webapp that screens the crypto market"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "^1.26.4"
flask = "^3.0.3"
flask-socketio = "^5.3.6"
websocket = "^0.2.1"
websocket-client = "^1.7.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
